requirements.txt
kucoin-futures-python==1.0.3
kucoin-python==1.0.7

